Move useAddEvent to own file
